[web/web-sys] Move gamepad_manager from backend
