+++ src/platform_impl/web/event_loop/global.rs
@@ -1,10 +1,11 @@
 use super::{backend, runner};
 use crate::event::{device, Event};
-use crate::platform_impl::platform::device::{gamepad, GamepadHandle};
+use crate::platform_impl::platform::device::{gamepad, gamepad_manager, GamepadHandle};
 use std::cell::RefCell;
 
 pub struct Emitter {
     window: RefCell<Option<backend::window::Shared>>,
+    gamepad_manager: gamepad_manager::Shared,
     gamepad_events: RefCell<Vec<(backend::gamepad::Gamepad, device::GamepadEvent)>>,
 }
 
@@ -12,6 +13,7 @@ impl Emitter {
     pub fn new() -> Self {
         Self {
             window: RefCell::new(None),
+            gamepad_manager: gamepad_manager::Shared::new(),
             gamepad_events: RefCell::new(Vec::new()),
         }
     }
@@ -24,8 +26,10 @@ impl Emitter {
             let shared = backend::window::Shared::create()?;
             let mut window = shared.0.borrow_mut();
 
+            let manager = self.gamepad_manager.clone().manager();
             let r = runner.clone();
             window.on_gamepad_connected(move |gamepad: backend::gamepad::Gamepad| {
+                let gamepad = manager.register(gamepad);
                 r.send_event(Event::GamepadEvent(
                     device::GamepadHandle(GamepadHandle {
                         id: gamepad.index() as i32,
@@ -35,8 +39,10 @@ impl Emitter {
                 ));
             });
 
+            let manager = self.gamepad_manager.clone().manager();
             let r = runner.clone();
             window.on_gamepad_disconnected(move |gamepad: backend::gamepad::Gamepad| {
+                let gamepad = manager.register(gamepad);
                 r.send_event(Event::GamepadEvent(
                     device::GamepadHandle(GamepadHandle {
                         id: gamepad.index() as i32,
@@ -54,13 +60,8 @@ impl Emitter {
 
     pub fn collect_gamepad_events(&self) -> Vec<(device::GamepadHandle, device::GamepadEvent)> {
         let mut gamepad_events = self.gamepad_events.borrow_mut();
-        match *self.window.borrow() {
-            Some(ref shared) => {
-                let window = shared.0.borrow();
-                window.collect_gamepad_events(&mut gamepad_events);
-            }
-            None => (),
-        }
+        self.gamepad_manager.manager().collect_events(&mut gamepad_events);
+        // backend::log(&format!("{:?}", gamepad_events).to_string());
 
         gamepad_events
             .drain(..)
